start refactoring of OSHEntities utility class, for #327, WIP
