+++ oshdb/src/test/java/org/heigit/ohsome/oshdb/osh/OSHRelationTest.java
@@ -142,93 +142,6 @@ public class OSHRelationTest {
     }
   }
 
-  @Test
-  public void testGetModificationTimestamps() throws IOException {
-    List<OSMNode> n1versions = new ArrayList<>();
-    n1versions.add(new OSMNode(123l, 2, new OSHDBTimestamp(2l), 12l, 0, new int[]{}, 0, 0));
-    n1versions.add(new OSMNode(123l, 1, new OSHDBTimestamp(1l), 11l, 0, new int[]{}, 0, 0));
-    OSHNode hnode1 = OSHNodeImpl.build(n1versions);
-    List<OSMNode> n2versions = new ArrayList<>();
-    n2versions.add(new OSMNode(124l, 4, new OSHDBTimestamp(12l), 24l, 0, new int[]{}, 0, 0));
-    n2versions.add(new OSMNode(124l, 3, new OSHDBTimestamp(9l), 23l, 0, new int[]{}, 0, 0));
-    n2versions.add(new OSMNode(124l, 2, new OSHDBTimestamp(4l), 22l, 0, new int[]{}, 0, 0));
-    n2versions.add(new OSMNode(124l, 1, new OSHDBTimestamp(3l), 21l, 0, new int[]{}, 0, 0));
-    OSHNode hnode2 = OSHNodeImpl.build(n2versions);
-    List<OSMNode> n3versions = new ArrayList<>();
-    n3versions.add(new OSMNode(125l, 3, new OSHDBTimestamp(11l), 34l, 0, new int[]{}, 0, 0));
-    n3versions.add(new OSMNode(125l, 2, new OSHDBTimestamp(6l), 32l, 0, new int[]{}, 0, 0));
-    n3versions.add(new OSMNode(125l, 1, new OSHDBTimestamp(1l), 31l, 0, new int[]{}, 0, 0));
-    OSHNode hnode3 = OSHNodeImpl.build(n3versions);
-
-    List<OSMWay> w1versions = new ArrayList<>();
-    w1versions.add(new OSMWay(1, 3, new OSHDBTimestamp(7l), 4445l, 23, new int[]{}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 0), new OSMMember(124, OSMType.NODE, 0)}));
-    w1versions.add(new OSMWay(1, 2, new OSHDBTimestamp(5l), 4444l, 23, new int[]{}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 0), new OSMMember(124, OSMType.NODE, 0), new OSMMember(125, OSMType.NODE, 0)}));
-    w1versions.add(new OSMWay(1, 1, new OSHDBTimestamp(4l), 4443l, 23, new int[]{}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 0), new OSMMember(124, OSMType.NODE, 0), new OSMMember(125, OSMType.NODE, 0)}));
-    OSHWay hway1 = OSHWayImpl.build(w1versions, Arrays.asList(hnode1, hnode2, hnode3));
-
-    List<OSMRelation> versions = new ArrayList<>();
-    versions.add(new OSMRelation(1, -4, new OSHDBTimestamp(20l), 10004l, 1, new int[]{}, new OSMMember[]{}));
-    versions.add(new OSMRelation(1, 3, new OSHDBTimestamp(10l), 10003l, 1, new int[]{1, 1, 2, 2}, new OSMMember[]{new OSMMember(1, OSMType.WAY, 0)}));
-    versions.add(new OSMRelation(1, 2, new OSHDBTimestamp(8l), 10002l, 1, new int[]{1, 1, 2, 2}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 1)}));
-    versions.add(new OSMRelation(1, 1, new OSHDBTimestamp(5l), 10001l, 1, new int[]{1, 1, 2, 2}, new OSMMember[]{new OSMMember(1, OSMType.WAY, 0)}));
-    OSHRelation hrelation = OSHRelationImpl.build(versions, Arrays.asList(hnode1, hnode2, hnode3), Arrays.asList(hway1));
-
-    List<OSHDBTimestamp> tss = OSHEntities.getModificationTimestamps(hrelation,false);
-    assertNotNull(tss);
-    assertEquals(4, tss.size());
-    assertEquals(5l, tss.get(0).getRawUnixTimestamp());
-    assertEquals(8l, tss.get(1).getRawUnixTimestamp());
-    assertEquals(10l, tss.get(2).getRawUnixTimestamp());
-    assertEquals(20l, tss.get(3).getRawUnixTimestamp());
-
-    tss = OSHEntities.getModificationTimestamps(hrelation,true);
-    assertNotNull(tss);
-    assertEquals(7, tss.size());
-    assertEquals(5l, tss.get(0).getRawUnixTimestamp());
-    assertEquals(6l, tss.get(1).getRawUnixTimestamp());
-    assertEquals(7l, tss.get(2).getRawUnixTimestamp());
-    assertEquals(8l, tss.get(3).getRawUnixTimestamp());
-    // timestamp 9 has to be missing because at the time, the way wasn't a member of the relation anymore
-    assertEquals(10l,  tss.get(4).getRawUnixTimestamp());
-    // timestamp 11 has to be missing because at the time, the node wasn't part of the way member of the relation anymore
-    assertEquals(12l, tss.get(5).getRawUnixTimestamp());
-    assertEquals(20l, tss.get(6).getRawUnixTimestamp());
-  }
-
-  @Test
-  public void testGetModificationTimestampsWithFilter() throws IOException {
-    List<OSMNode> n1versions = new ArrayList<>();
-    n1versions.add(new OSMNode(123l, 7, new OSHDBTimestamp(17l), 17l, 0, new int[]{}, 0, 0));
-    n1versions.add(new OSMNode(123l, 6, new OSHDBTimestamp(6l), 16l, 0, new int[]{}, 0, 0));
-    n1versions.add(new OSMNode(123l, 5, new OSHDBTimestamp(5l), 15l, 0, new int[]{}, 0, 0));
-    n1versions.add(new OSMNode(123l, 4, new OSHDBTimestamp(4l), 14l, 0, new int[]{}, 0, 0));
-    n1versions.add(new OSMNode(123l, 3, new OSHDBTimestamp(3l), 13l, 0, new int[]{}, 0, 0));
-    n1versions.add(new OSMNode(123l, 2, new OSHDBTimestamp(2l), 12l, 0, new int[]{}, 0, 0));
-    n1versions.add(new OSMNode(123l, 1, new OSHDBTimestamp(1l), 11l, 0, new int[]{}, 0, 0));
-    OSHNode hnode1 = OSHNodeImpl.build(n1versions);
-
-    List<OSMRelation> versions = new ArrayList<>();
-    versions.add(new OSMRelation(1, -4, new OSHDBTimestamp(6l), 10004l, 1, new int[]{}, new OSMMember[]{}));
-    versions.add(new OSMRelation(1, 3, new OSHDBTimestamp(5l), 10003l, 1, new int[]{}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 0)}));
-    versions.add(new OSMRelation(1, 2, new OSHDBTimestamp(3l), 10002l, 1, new int[]{}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 1)}));
-    versions.add(new OSMRelation(1, 1, new OSHDBTimestamp(1l), 10001l, 1, new int[]{}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 0)}));
-    OSHRelation hrelation = OSHRelationImpl.build(versions, Arrays.asList(hnode1), Arrays.asList());
-
-    List<OSHDBTimestamp> tss = OSHEntities.getModificationTimestamps(
-        hrelation,
-        entity -> entity.getVersion() != 2
-    );
-    assertNotNull(tss);
-    assertEquals(5, tss.size());
-    assertEquals(1l, tss.get(0).getRawUnixTimestamp());
-    assertEquals(2l, tss.get(1).getRawUnixTimestamp());
-    assertEquals(3l, tss.get(2).getRawUnixTimestamp());
-    // ts 4 missing since entity filter doesn't match then
-    assertEquals(5l, tss.get(3).getRawUnixTimestamp());
-    assertEquals(6l, tss.get(4).getRawUnixTimestamp());
-
-  }
-
   static OSHNode buildHOSMNode(List<OSMNode> versions) {
     try {
       return OSHNodeImpl.build(versions);
@@ -256,39 +169,5 @@ public class OSHRelationTest {
     String expResult = "OSHRelation ID:300 Vmax:+1+ Creation:3333 BBox:(8.680973,49.409498),(8.680973,49.409498)";
     String result = instance.toString();
     assertEquals(expResult, result);
-  }
-
-  @Test
-  public void testIssue325() throws IOException {
-    // tests that the bug reported in https://github.com/GIScience/oshdb/issues/325 is fixed:
-    // relations referencing redacted ways caused a crash in the OSHEntities utility class
-    // when calculating the relation's modification timestamps
-    OSHNode hnode1 = OSHNodeImpl.build(new ArrayList<>(List.of(
-        new OSMNode(123L, 2, new OSHDBTimestamp(2L), 2L, 0, new int[]{}, 0, 0),
-        new OSMNode(123L, 1, new OSHDBTimestamp(1L), 1L, 0, new int[]{}, 0, 0))));
-    OSHNode hnode2 = OSHNodeImpl.build(new ArrayList<>(List.of(
-        new OSMNode(124L, 2, new OSHDBTimestamp(2L), 2L, 0, new int[]{}, 0, 0),
-        new OSMNode(124L, 1, new OSHDBTimestamp(1L), 1L, 0, new int[]{}, 0, 0))));
-
-    OSHWay hway1 = OSHWayImpl.build(new ArrayList<>(List.of(
-        new OSMWay(1, 1, new OSHDBTimestamp(1L), 1L, 0, new int[]{}, new OSMMember[]{
-            new OSMMember(123L, OSMType.NODE, 0),
-            new OSMMember(124L, OSMType.NODE, 0)})
-    )), List.of(hnode1, hnode2));
-
-    OSHWay hway2 = OSHWayImpl.build(new ArrayList<>(List.of(
-        new OSMWay(2L, -4, new OSHDBTimestamp(9L), 9L, 9, new int[]{}, new OSMMember[]{})
-    )), Collections.emptyList());
-
-    OSHRelation hrelation = OSHRelationImpl.build(new ArrayList<>(List.of(
-        new OSMRelation(1L, 2, new OSHDBTimestamp(8L), 8L, 8, new int[]{1, 1, 2, 2}, new OSMMember[]{
-            new OSMMember(1L, OSMType.WAY, 0),
-            new OSMMember(2L, OSMType.WAY, 0)}),
-        new OSMRelation(1L, 1, new OSHDBTimestamp(1L), 1L, 0, new int[]{1, 1, 2, 2}, new OSMMember[]{
-            new OSMMember(1L, OSMType.WAY, 0)})
-    )), List.of(hnode1, hnode2), List.of(hway1, hway2));
-
-    List<OSHDBTimestamp> tss = OSHEntities.getModificationTimestamps(hrelation, true);
-    assertNotNull(tss);
   }
 }
